Update Limits.java master

+++ src/main/java/com/in28minutes/microservices/limitsservice/bean/Limits.java
@@ -4,7 +4,7 @@ public class Limits {
 	private int minimum;
 	private int maximum;
 	
-	private int longitude;
+	private long longitude;
 	private int latitude;
 
 	public Limits() {
@@ -12,6 +12,7 @@ public class Limits {
 	}
 
 	public Limits(int minimum, int maximum) {
+		//we will use this
 		super();
 		this.minimum = minimum;
 		this.maximum = maximum;
@@ -33,11 +34,11 @@ public class Limits {
 		this.maximum = maximum;
 	}
 
-	public int getLongitude() {
+	public long getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(int longitude) {
+	public void setLongitude(long longitude) {
 		this.longitude = longitude;
 	}
